_header.jsリファクタリング
